% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydyn.R
\name{confeqn}
\alias{confeqn}
\title{Configure an equation for phydyn analysis}
\usage{
confeqn(
  eqn,
  type = c("birth", "death", "migration", "nondeme", "definition"),
  origin = "",
  destination = ""
)
}
\arguments{
\item{eqn}{A character string specifying the rate of a given event (birth, death, etc., see *type*)}

\item{type}{A character string specifying if the given equation gives the rate of birth, death, or migration between demes. Additionally, *eqn* can specify the rate of change of a `nondeme` variable, which is a dynamic variable that is not a deme.}

\item{origin}{If type is `birth`, `death`, `migration` or `nondeme`, this specifies the name of the deme or variable corresponding to the rate specified in `eqn`}

\item{destination}{If type is `birth` or `migration` this is a required name of a second deme that is part of the rate specified in `eqn`. For example, if `birth`, the `eqn` specifies the rate that units in deme `origin` generates a copy in `destination`. If `migration`, the `eqn` specifies the rate that units in deme `origin` changes state to deme `destination`.}
}
\value{
An equation of class `phydynequation`
}
\description{
This function allows for the configuration of a model element, including rates of birth, death, migration between demes, auxliary dynamic state variables (nondeme), and definition of derived quantities. Some model elements (e.g. birth) must specify an *origin* and *destination* since births can happen within or between demes. Other elements (e.g. death) only need to specify *origin*.
}
\details{
For information about PhyDyn model syntax and supported mathematical syntax and operations, see \url{https://github.com/mrc-ide/PhyDyn/wiki}.
}
