% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydyn.R
\name{confparm}
\alias{confparm}
\title{Configure a numeric parameter for PhyDyn analysis}
\usage{
confparm(
  name = "",
  initial_condition_parameter = FALSE,
  initial = NA,
  estimate = TRUE,
  prior = c("lognormal", "normal", "exponential", "beta"),
  operator = c("realrw", "scale"),
  lower = -Inf,
  upper = Inf,
  weight = 1,
  ...
)
}
\arguments{
\item{name}{The name of the parameter}

\item{initial_condition_parameter}{TRUE if this parameter specifies the initial value of a dynamic variable, such as the size of a deme specified in one of the PhyDyn equations. In that case, the parameter name should be the same as the deme name.}

\item{initial}{The starting value (numeric) for this parameter in MCMC sampling.}

\item{estimate}{If TRUE, this parameter will be estimated by MCMC sampling. If FALSE, the value is fixed to *initial* and there is no need to specify priors or operators.}

\item{prior}{A character string specifying the prior distribution. See details.}

\item{operator}{A character string specifying the method of MCMC moves for this parameter. See details.}

\item{lower}{Numeric lower bound of this parameter.}

\item{upper}{Numeric upper bound of this parameter.}

\item{...}{Additional arguments may be required depending on the prior and operator. See details.}
}
\value{
A `phydynparameter` to be passed to `config_phydyn`
}
\description{
This sets the priors, operators (for MCMC sampling), upper and lower bounds, and initial conditions for a PhyDyn model parameter.
The names of parameters must correspond to those used in PhyDyn model equations. It is important to carefully specify priors and operators for all parameters since these choices can have a large influence on estimated posteriors and the efficiency of MCMC sampling.
}
\details{
Supported priors are listed below. Note these may require additional arguments passed to ... 

\itemize{
	\item  normal: This has required arguments `mean` and `sigma`
	\item  lognormal: This has required arguments `M` and `S`
	\item  exponential: This has required argument `mean`
	\item  beta: This has required arguments `alpha` and `beta`. 
}

Supported operators which specify MCMC moves are listed below. Note that these may require additional arguments passed to ... 

\itemize{
	\item  realrw: A random walk move. Should specify `windowSize` which is the width of the numeric interval used for the MCMC move. If not set, this will be set to 10% of the initial parameter value.  Optionally pass `useGaussian` set to "true"(default) or "false" to specify if normal or uniform moves should be used. 
	\item  scale: This will inflate or shrink the parameter. Useful for parameters > 0 such as population sizes or rates. Should specify `scaleFactor` (default=0.5) which is the magnitude of the rescaling move. 
}

Passing `weight` (default=1) to ... will influence how frequently MCMC moves for this parameter are performed.
}
